Fixed things for compatibility and added warnings for gcc and cmake
Added warning flags for clang and gcc. Rebuild configuration for it to take effect
Added scripts to build and clean all projects with cmake
Fixed things in each project

AEMisc:
Added another type for ace::math::equals() for better comparisons (AEMath.hpp)
Fixed preliminary checks in ace::math::intPow() (AEMath.hpp)
Fixed raw comparison of floats in ace::math::intPow() (AEMath.hpp)
ace::math::_getTrigTableIndex() is more readable, aside from constexpr branch (AETrigLookup.hpp)
Changed typedefs of ullint, llint, sint, uint, to use the primitive and *explicit* types (not implementation-defined int64_t, etc) for compatibility (AETypedefs.hpp)
Fixed *date functions having a c-style zero-initialiser on a tm tstruct instead of C++ style (AEUtils.hpp)

AEModuleBase:
Deleted implicit copy-assignment operator
Removed the *this = val from the copy-constructor
Added COPYCONSTRUCT_MODULE() macro to call the __AEModuleBase<> copy constructor, if the class is explicitly defining it

AEVector:
Fixed comparison operator not using ace::math::equals()

AEFrame:
Added an explicit copy-constructor and copy-assignment operator

AETimer:
Fixed the copy-constructor not invoking the __AEModuleBase<> copy constructor

AELogEntry:
Changed setting the m_ullCurrentEntry from -1 to ULLINT_MAX to remove a compiler warning (they do the same thing anyway)

+++ AEMisc/include/AETypedefs.hpp
@@ -25,16 +25,16 @@ typedef uint8_t ucint;
 typedef int8_t cint;
 
 /// A shortcut for big unsigned (long long) integer (64bit)
-typedef uint64_t ullint;
+typedef unsigned long long ullint;
 
 /// A shortcut for big signed (long long) integer (64bit)
-typedef int64_t llint;
+typedef long long llint;
 
 /// A shortcut for a just signed int (32bit)
-typedef int32_t sint;
+typedef signed int sint;
 
 /// A shortcut for a just unsigned int (32bit)
-typedef uint32_t uint;
+typedef unsigned int uint;
 
 /// A shortcut for a just unsigned short (at least 16 bit)
 typedef unsigned short ushort;
